Rename InteractiveFlowStatus to OAuth2AuthorizeInteractiveFlowStatus (#276)
Move the status out of the generic business.model.flow package into business.model.flow.auth and rename it OAuth2AuthorizeInteractiveFlowStatus: it is computed by and exclusive to the OAuth2 authorize purpose handler, so it does not belong to the generic interactive-flow model.
